--- frontend/src/app/(auth)/Login.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-import SignInSide from "../../components/MUI components/CustomSignIn/SignInSide";
-
-const Login = () => {
-  return (
-    <div>
-      <p>hello</p>
-      <SignInSide  disableCustomTheme={true}/>
-    </div>
-  );
-};
-
-export default Login;
