--- frontend/src/components/MUI components/CustomSignIn/README.md
@@ -1,15 +1,0 @@
-# Sign-in side template
-
-## Usage
-
-<!-- #default-branch-switch -->
-
-1. Copy these folders (`sign-in-side` and `shared-theme`) into your project, or one of the [example projects](https://github.com/mui/material-ui/tree/master/examples).
-2. Make sure your project has the required dependencies: @mui/material, @mui/icons-material, @emotion/styled, @emotion/react.
-3. Import and use the `SignInSide` component.
-
-## Demo
-
-<!-- #default-branch-switch -->
-
-View the demo at https://mui.com/material-ui/getting-started/templates/sign-in-side/.
